// src/lib/types/JobPreferences.d.ts
export interface JobPreferences {
	answers: {
		city: string;
		workType: string;
		employmentType: string;
		travelDistance: string;
		salary: string;
		interest1: string;
		interest2: string;
		interest3: string;
		interest4: string;
		interest5: string;
		strength1: string;
		strength2: string;
		strength3: string;
		goal1: string;
		goal2: string;
	};
}